pyproject.toml
[project]
name = "django-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-browser-reload>=1.18.0",
    "django-cleanup>=9.0.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "waitress>=3.0.2",
    "whitenoise>=6.9.0",
]
